MOB-2902 [Quotes V2 -> Quotes - remove deprecated properties] (#37)
* QuoteV2 -> Quote

* mac os support

* lower deployment target mac os

* 10.15 support

* conditional import UIKit for macos

* conditional import UIKit

* conditional import UIKit

* update version

* remove v2 refs

* remove v2 referance
