+++ KarhooSDK/Api/Response/QuoteList.swift
@@ -9,9 +9,6 @@ typealias QuoteListKeys = QuoteList.CodingKeys
 
 public struct QuoteList: KarhooCodableModel {
 
-    @available(*, deprecated, message: "use quotes (QuotesV2)")
-    public let quoteItems: [Quote]
-
     public let quotes: [Quote]
     public let listId: String
     public let status: String
@@ -24,7 +21,6 @@ public struct QuoteList: KarhooCodableModel {
                   status: String = "",
                   validity: Int = 0,
                   availability: Availability = Availability()) {
-        self.quoteItems = quoteItems
         self.listId = listId
         self.status = status
         self.validity = validity
@@ -33,7 +29,6 @@ public struct QuoteList: KarhooCodableModel {
     }
 
     enum CodingKeys: String, CodingKey {
-        case quoteItems = "quote_items"
         case quotes
         case listId = "id"
         case status
@@ -43,7 +38,6 @@ public struct QuoteList: KarhooCodableModel {
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.quoteItems = (try? container.decode([Quote].self, forKey: .quoteItems)) ?? []
         self.quotes = (try? container.decode([Quote].self, forKey: .quotes)) ?? []
         self.listId = try container.decode(String.self, forKey: .listId)
         self.status = (try? container.decode(String.self, forKey: .status)) ?? ""
@@ -54,7 +48,6 @@ public struct QuoteList: KarhooCodableModel {
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(quotes, forKey: .quotes)
-        try container.encode(quoteItems, forKey: .quoteItems)
         try container.encode(listId, forKey: .listId)
         try container.encode(status, forKey: .status)
         try container.encode(validity, forKey: .validity)
