--- KarhooSDK/Service/Availability/AvailabilityInteractor/AvailabilityInteractor.swift
@@ -1,11 +1,0 @@
-//
-// 
-// Copyright © 2020 Karhoo. All rights reserved.
-//
-
-import Foundation
-
-public protocol AvailabilityInteractor: KarhooExecutable {
-
-    func set(availabilitySearch: AvailabilitySearch)
-}
