+++ KarhooSDK/Service/Quote/QuoteService.swift
@@ -8,8 +8,6 @@
 
 import Foundation
 
-public protocol QuoteService {
+public protocol QuoteService {    
     func quotes(quoteSearch: QuoteSearch) -> PollCall<Quotes>
-    
-    func quotesV2(quoteSearch: QuoteSearch) -> PollCall<Quotes>
 }
