+++ KarhooSDKTests/TestCases/Service/Quote/Interactor/KarooQuoteInteractorSpec.swift
@@ -2,8 +2,8 @@
 //  KarhooQuoteInteractorSpec.swift
 //  KarhooSDKTests
 //
-//  
-//  Copyright © 2020 Karhoo. All rights reserved.
+//  Created by Jeevan Thandi on 14/07/2020.
+//  Copyright © 2020 Flit Technologies Ltd. All rights reserved.
 //
 
 import Foundation
@@ -14,47 +14,38 @@ import XCTest
 final class KarooQuoteInteractorSpec: XCTestCase {
 
     private var testObject: KarhooQuoteInteractor!
-    private var mockAvailabilityRequest: MockRequestSender!
     private var mockQuoteListIdRequest: MockRequestSender!
     private var mockQuotesRequest: MockRequestSender!
 
     private lazy var mockQuoteSearch: QuoteSearch = {
-        return QuoteSearch(origin: LocationInfoMock().set(placeId: "originPlaceId").build(),
-                           destination: LocationInfoMock().set(placeId: "destinationPlaceId").build(),
+        return QuoteSearch(origin: LocationInfoMock().set(position: Position(latitude: 0.1, longitude: 0.2)).build(),
+                           destination: LocationInfoMock().set(position: Position(latitude: 0.2, longitude: 0.3)).build(),
                            dateScheduled: Date())
-    }()
-
-    private lazy var mockCategories: Categories = {
-        return CategoriesMock().set(categories: ["foo"]).build()
     }()
 
     override func setUp() {
         super.setUp()
 
-        mockAvailabilityRequest = MockRequestSender()
         mockQuoteListIdRequest = MockRequestSender()
         mockQuotesRequest = MockRequestSender()
 
-        testObject = KarhooQuoteInteractor(availabilityRequest: mockAvailabilityRequest,
-                                           quoteListIdRequest: mockQuoteListIdRequest,
-                                           quotesRequest: mockQuotesRequest)
+        testObject = KarhooQuoteInteractor(quoteListIdRequest: mockQuoteListIdRequest,
+                                             quotesRequest: mockQuotesRequest)
 
         testObject.set(quoteSearch: mockQuoteSearch)
     }
 
     /**
-      * When: Calling execute
-      * And: No QuoteSearch is set on the interactor
-      * And: No requests should be made
-      */
+     * When: Calling execute
+     * And: No QuoteSearch is set on the interactor
+     * And: No requests should be made
+     */
     func testNoQuoteSearchSet() {
-        let noQuoteSearchSetInteractor = KarhooQuoteInteractor(availabilityRequest: mockAvailabilityRequest,
-                                                               quoteListIdRequest: mockQuoteListIdRequest,
-                                                               quotesRequest: mockQuotesRequest)
+        let noQuoteSearchSetInteractor = KarhooQuoteInteractor(quoteListIdRequest: mockQuoteListIdRequest,
+                                                                 quotesRequest: mockQuotesRequest)
 
         noQuoteSearchSetInteractor.execute(callback: { (_:Result<Quotes>) in })
 
-        XCTAssertFalse(mockAvailabilityRequest.requestAndDecodeCalled)
         XCTAssertFalse(mockQuoteListIdRequest.requestAndDecodeCalled)
         XCTAssertFalse(mockQuotesRequest.requestAndDecodeCalled)
     }
@@ -65,42 +56,42 @@ final class KarooQuoteInteractorSpec: XCTestCase {
      * Then: The expected pickuptime string, localised in the correct format
      *       Should be propogated to the availabilityRequest
      */
-    func testDateScheduledFormatInAvailabilityRequest() {
+    func testDateScheduledFormatInQuoteRequest() {
         let testDate = Date(timeIntervalSince1970: 1605195000) // UTC: 11/12/2020-15:30
 
         let testQuoteSearch = QuoteSearch(origin: LocationInfoMock().set(timeZoneIdentifier: "America/New_York")
-                                                                    .set(placeId: "origin_id")
-                                                                    .build(),
+            .set(placeId: "origin_id")
+            .build(),
                                           destination: LocationInfoMock().set(placeId: "destination_id")
-                                                                         .build(),
+                                            .build(),
                                           dateScheduled: testDate)
 
         testObject.set(quoteSearch: testQuoteSearch)
         testObject.execute(callback: { (_: Result<Quotes>) in })
 
-        let availabilityRequestPayload = mockAvailabilityRequest.payloadSet as? AvailabilitySearch
+        let mockQuoteSearchPayload = mockQuoteListIdRequest.payloadSet as? QuoteRequest
 
-        XCTAssertEqual("2020-11-12T10:30", availabilityRequestPayload?.dateScheduled)
+        XCTAssertEqual("2020-11-12T10:30", mockQuoteSearchPayload?.dateScheduled)
     }
 
     /**
      * When: Cancelling Quote Search
-     * Then: Availability and QuoteListId request should be cancelled
+     * Then: QuoteListId request should be cancelled
      * And: QuotePoller should be stopped
      */
     func testCancelQuoteSearch() {
         testObject.cancel()
 
-        XCTAssertTrue(mockAvailabilityRequest.cancelNetworkRequestCalled)
+        XCTAssertTrue(mockQuotesRequest.cancelNetworkRequestCalled)
         XCTAssertTrue(mockQuoteListIdRequest.cancelNetworkRequestCalled)
     }
 
     /**
-      * Given: Calling execute
-      * When: the interactor does not hold a quote list id
-      * Then: request for availability should be made
-      * And: Quote request should not be called
-      */
+     * Given: Calling execute
+     * When: the interactor does not hold a quote list id
+     * Then: request for quote list id should be made
+     * And: Quote request should not be called
+     */
     func testNoQuoteListId() {
         testObject.execute(callback: { (_: Result<Quotes>) in })
 
@@ -112,64 +103,28 @@ final class KarooQuoteInteractorSpec: XCTestCase {
             dateScheduled = dateFormatter.toString(from: quoteSearchDate)
         }
 
-        let expectedPayload = AvailabilitySearch(origin: mockQuoteSearch.origin.placeId,
-                                                 destination: mockQuoteSearch.destination.placeId,
-                                                 dateScheduled: dateScheduled)
+        let expectedOrigin = QuoteRequestPoint(latitude: "\(mockQuoteSearch.origin.position.latitude)",
+                                               longitude: "\(mockQuoteSearch.origin.position.longitude)",
+                                               displayAddress: mockQuoteSearch.origin.address.displayAddress)
 
-        mockAvailabilityRequest.assertRequestSendAndDecoded(endpoint: .availability,
-                                                            method: .get,
-                                                            payload: expectedPayload)
+        let expectedDestination = QuoteRequestPoint(latitude: "\(mockQuoteSearch.destination.position.latitude)",
+                                                    longitude: "\(mockQuoteSearch.destination.position.longitude)",
+                                                    displayAddress: mockQuoteSearch.destination.address.displayAddress)
+
+        let expectedPayload = QuoteRequest(origin: expectedOrigin,
+                                           destination: expectedDestination,
+                                           dateScheduled: dateScheduled)
+
+        mockQuoteListIdRequest.assertRequestSendAndDecoded(endpoint: .quoteListId,
+                                                           method: .post,
+                                                           payload: expectedPayload)
 
         XCTAssertFalse(mockQuotesRequest.requestAndDecodeCalled)
     }
 
     /**
-      * Given: Calling execute
-      * When: The interactor does not hold a quote list id
-      * And: the request for availability succeeds
-      * Then: quote list id request should be made
-      */
-    func testRequestAndHandleAvailabilityRequestSuccess() {
-        testObject.execute(callback: { (_: Result<Quotes>) in })
-
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
-
-        let availabilityRequestPayload = mockAvailabilityRequest.payloadSet as? AvailabilitySearch
-
-        let expectedPayload = QuoteListIdRequestPayload(origin: mockQuoteSearch.origin.placeId,
-                                                        destination: mockQuoteSearch.destination.placeId,
-                                                        dateScheduled: availabilityRequestPayload?.dateScheduled)
-
-        mockQuoteListIdRequest.assertRequestSendAndDecoded(endpoint: .quoteListId,
-                                                           method: .post,
-                                                           payload: expectedPayload)
-    }
-
-    /**
      * Given: Calling execute
      * When: The interactor does not hold a quote list id
-     * And: the request for availability fails
-     * Then: quote list id request should be made
-     */
-    func testRequestAndHandleAvailabilityRequestFails() {
-        testObject.execute(callback: { (_: Result<Quotes>) in })
-
-        mockAvailabilityRequest.triggerFail(error: TestUtil.getRandomError())
-        let availabilityRequestPayload = mockAvailabilityRequest.payloadSet as? AvailabilitySearch
-
-        let expectedPayload = QuoteListIdRequestPayload(origin: mockQuoteSearch.origin.placeId,
-                                                        destination: mockQuoteSearch.destination.placeId,
-                                                        dateScheduled: availabilityRequestPayload?.dateScheduled)
-
-        mockQuoteListIdRequest.assertRequestSendAndDecoded(endpoint: .quoteListId,
-                                                           method: .post,
-                                                           payload: expectedPayload)
-    }
-
-    /**
-     * Given: Calling execute
-     * When: The interactor does not hold a quote list id
-     * And: the request for availability completes
      * And: quote list id request fails
      * Then: Callback should be propogated
      */
@@ -178,10 +133,9 @@ final class KarooQuoteInteractorSpec: XCTestCase {
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerFail(error: expectedError)
 
-        XCTAssert(expectedError.equals(result!.errorValue()!))
+        XCTAssertNotNil(expectedError.equals(result?.errorValue()))
     }
 
     /**
@@ -194,7 +148,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
     func testQuoteListIdRequestSucceeds() {
         testObject.execute(callback: { (_: Result<Quotes>) in })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
         mockQuotesRequest.assertRequestSendAndDecoded(endpoint: .quotes(identifier: "some"),
@@ -205,7 +158,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
     /**
      * Given: Calling execute
      * When: The interactor does not hold a quote list id
-     * And: the request for availability completes
      * And: quote list id request succeeds
      * And: quote request succeeds
      * Then: callback should propogate expected value
@@ -214,7 +166,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
         let expectedQuote = QuoteMock().set(quoteId: "foo").set(categoryName: "foo").build()
@@ -226,13 +177,12 @@ final class KarooQuoteInteractorSpec: XCTestCase {
 
         mockQuotesRequest.triggerSuccessWithDecoded(value: expectedResult)
 
-        XCTAssertEqual(expectedQuote.quoteId, result!.successValue()!.quotes(for: "foo")[0].quoteId)
+        XCTAssertEqual(expectedQuote.id, result?.successValue()?.all[0].id)
     }
 
     /**
      * Given: Calling execute
      * When: The interactor does not hold a quote list id
-     * And: the request for availability completes
      * And: quote list id request succeeds
      * And: quote request fails
      * Then: quote list id request should be made again
@@ -243,18 +193,16 @@ final class KarooQuoteInteractorSpec: XCTestCase {
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
         mockQuotesRequest.triggerFail(error: expectedError)
 
-        XCTAssert(expectedError.equals(result!.errorValue()))
+        XCTAssert(expectedError.equals(result?.errorValue()))
     }
 
     /**
      * Given: Calling execute
      * When: The interactor does not hold a quote list id
-     * And: the request for availability completes
      * And: quote list id request succeeds
      * And: quote request fails due to K3001
      * Then: quote list id request should be made again
@@ -264,14 +212,12 @@ final class KarooQuoteInteractorSpec: XCTestCase {
 
         testObject.execute(callback: { (_: Result<Quotes>) in })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
         mockQuoteListIdRequest.requestAndDecodeCalled = false
 
         mockQuotesRequest.triggerFail(error: expectedError)
 
-        XCTAssertTrue(mockAvailabilityRequest.cancelNetworkRequestCalled)
         XCTAssertTrue(mockQuoteListIdRequest.cancelNetworkRequestCalled)
         XCTAssertTrue(mockQuoteListIdRequest.requestAndDecodeCalled)
     }
@@ -279,7 +225,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
     /**
      * Given: Calling execute
      * When: the interactor holds a quote list id
-     * Then: request for availability should not be made
      * Then: request for quote list id should not be made
      * And: Quote request should be made
      * And: Result should propogate in the callback
@@ -287,22 +232,18 @@ final class KarooQuoteInteractorSpec: XCTestCase {
     func testQuoteListIdAlreadyPresent() {
         testObject.execute(callback: { (_: Result<Quotes>) in })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
-        XCTAssertTrue(mockAvailabilityRequest.requestAndDecodeCalled)
         XCTAssertTrue(mockQuoteListIdRequest.requestAndDecodeCalled)
         XCTAssertTrue(mockQuotesRequest.requestAndDecodeCalled)
 
-        // reset
-        mockAvailabilityRequest.requestAndDecodeCalled = false
+
         mockQuoteListIdRequest.requestAndDecodeCalled = false
         mockQuotesRequest.requestAndDecodeCalled = false
 
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        XCTAssertFalse(mockAvailabilityRequest.requestAndDecodeCalled)
         XCTAssertFalse(mockQuoteListIdRequest.requestAndDecodeCalled)
 
         XCTAssertTrue(mockQuotesRequest.requestAndDecodeCalled)
@@ -316,25 +257,23 @@ final class KarooQuoteInteractorSpec: XCTestCase {
 
         mockQuotesRequest.triggerSuccessWithDecoded(value: expectedResult)
 
-        XCTAssertEqual(expectedQuote.quoteId, result!.successValue()!.quotes(for: "foo")[0].quoteId)
+        XCTAssertEqual(expectedQuote.id, result?.successValue()?.all[0].id)
     }
 
     /**
      * Given: Calling execute
      * When: The interactor does not hold a quote list id
-     * And: the request for availability completes
      * And: quote list id request succeeds
-     * And: quote request succeeds but contains a quote with an ETA greater than 20 mins
+     * And: quote request succeeds but contains a quote with an ETA greater than 30 mins
      * Then: callback should propogate empty quotes list
      */
     func testQuoteWithGreaterThan20MinEtaIsFiltered() {
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 100))
 
-        let expectedQuote = QuoteMock().set(quoteId: "foo").set(categoryName: "foo").set(qtaHighMinutes: 21).build()
+        let expectedQuote = QuoteMock().set(quoteId: "foo").set(categoryName: "foo").set(qtaHighMinutes: 31).build()
 
         let expectedResult = QuoteListMock().set(quoteListId: "some_id")
             .set(status: "some_status")
@@ -348,7 +287,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
 
     /**
      * Given: Calling execute
-     * And: the request for availability completes
      * And: quote list id request succeeds but will expire within 10 seconds
      * Then: callback should not be called and request new QuoteListId
      * When: quote list returns QuoteList with validity above 10
@@ -358,7 +296,6 @@ final class KarooQuoteInteractorSpec: XCTestCase {
         var result: Result<Quotes>?
         testObject.execute(callback: { result = $0 })
 
-        mockAvailabilityRequest.triggerSuccessWithDecoded(value: mockCategories)
         mockQuoteListIdRequest.triggerSuccessWithDecoded(value: QuoteListId(identifier: "some", validityTime: 5))
 
         let expectedQuote = QuoteMock().set(quoteId: "foo").set(categoryName: "foo").build()
@@ -382,6 +319,7 @@ final class KarooQuoteInteractorSpec: XCTestCase {
         mockQuotesRequest.triggerSuccessWithDecoded(value: validQuoteList)
 
         XCTAssertNotNil(result)
-        XCTAssertTrue(result!.isSuccess())
+        XCTAssert(result?.isSuccess() == true)
     }
+
 }
