+++ KarhooSDKTests/TestCases/Service/Quote/KarhooQuoteServiceSpec.swift
@@ -14,8 +14,7 @@ import XCTest
 final class KarhooQuoteServiceSpec: XCTestCase {
 
     private var testobject: KarhooQuoteService!
-    private var mockQuoteInteractor: MockQuoteInteractor!
-    private var mockQuoteInteractorV2 = MockQuoteInteractorV2()
+    private var mockQuoteInteractor = MockQuoteInteractor()
 
     private let mockQuoteSearch: QuoteSearch = QuoteSearch(origin: LocationInfoMock()
                                                                    .set(placeId: "originPlaceId")
@@ -36,7 +35,7 @@ final class KarhooQuoteServiceSpec: XCTestCase {
 
         mockQuoteInteractor = MockQuoteInteractor()
 
-        testobject = KarhooQuoteService(quoteInteractor: mockQuoteInteractor, quoteV2Interactor: mockQuoteInteractorV2)
+        testobject = KarhooQuoteService(quoteInteractor: mockQuoteInteractor)
     }
 
     /**
@@ -51,7 +50,7 @@ final class KarhooQuoteServiceSpec: XCTestCase {
 
         mockQuoteInteractor.triggerSuccess(result: mockQuotesResult)
 
-        XCTAssertEqual("success-quote", result?.successValue()?.quotes(for: "foo")[0].quoteId)
+        XCTAssertEqual("success-quotev2", result?.successValue()?.quotes(for: "foo")[0].fleet.id)
     }
 
     /**
@@ -66,37 +65,6 @@ final class KarhooQuoteServiceSpec: XCTestCase {
 
         let expectedError = TestUtil.getRandomError()
         mockQuoteInteractor.triggerFail(error: expectedError)
-
-        XCTAssert(expectedError.equals(result?.errorValue()))
-    }
-    
-    /**
-      * When: QuoteV2 search succeeds
-      * Then: callback should be executed with expected value
-      */
-    func testQuoteV2SearchSucces() {
-        let pollCall = testobject.quotesV2(quoteSearch: mockQuoteSearch)
-
-        var result: Result<Quotes>?
-        pollCall.execute(callback: { result = $0 })
-
-        mockQuoteInteractorV2.triggerSuccess(result: mockQuotesResult)
-
-        XCTAssertEqual("success-quotev2", result?.successValue()?.quotes(for: "foo")[0].fleet.id)
-    }
-
-    /**
-     * When: Quote search fails
-     * Then: callback should be executed with expected value
-     */
-    func testQuoteV2SearchFails() {
-        let pollCall = testobject.quotesV2(quoteSearch: mockQuoteSearch)
-
-        var result: Result<Quotes>?
-        pollCall.execute(callback: { result = $0 })
-
-        let expectedError = TestUtil.getRandomError()
-        mockQuoteInteractorV2.triggerFail(error: expectedError)
 
         XCTAssert(expectedError.equals(result?.errorValue()))
     }
